Revamping the LeetCodeFiles repo
Decided to go ahead and organize this thing because I've been so committed to it. In here now are some powershell files that allow for consistent naming conventions, and PS1 files that allow for me to be able to go ahead and create all the files I need for the month. This didn't take that long, but I hope my <10 followers enjoy!